+++ src/main/kotlin/swyg/hollang/entity/Test.kt
@@ -7,8 +7,10 @@ import swyg.hollang.entity.common.BaseTimeEntity
 
 @Entity
 class Test private constructor(
+
     @Column(name = "version", unique = true, updatable = false, nullable = false)
-    val version: Long
+    val version: Int
+
 ) : BaseTimeEntity() {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -18,7 +20,7 @@ class Test private constructor(
     @OneToMany(mappedBy = "test", fetch = LAZY, cascade = [ALL], orphanRemoval = true)
     val questions: MutableSet<Question> = mutableSetOf()
 
-    constructor(version: Long, questions: MutableSet<Question>) : this(version) {
+    constructor(version: Int, questions: MutableSet<Question>) : this(version) {
         this.questions.addAll(questions)
         questions.forEach { question ->
             question.test = question.test ?: this
